Reorganize data structure on src repo
